+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Dedicated error message for ternary type mismatch. https://github.com/rescript-lang/rescript/pull/7804
 - Dedicated error message for passing a braced ident to something expected to be a record. https://github.com/rescript-lang/rescript/pull/7806
 - Hint about partial application when missing required argument in function call. https://github.com/rescript-lang/rescript/pull/7807
+- More autocomplete improvements involving modules and module types. https://github.com/rescript-lang/rescript/pull/7795
 
 #### :house: Internal